# [PHP] Use IDL namer in generator and add prefixed size, clear, and createBytesVector in builder