--- a/src/bfbs_gen_nim.cpp
+++ b/src/bfbs_gen_nim.cpp
@@ -68,6 +68,7 @@ Namer::Config NimDefaultConfig() {
            /*namespace_seperator=*/"/",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kKeep,