--- a/src/idl_gen_dart.cpp
+++ b/src/idl_gen_dart.cpp
@@ -46,6 +46,7 @@ static Namer::Config DartDefaultConfig() {
            /*namespace_seperator=*/".",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"$",
            /*keyword_suffix=*/"",
            /*filenames=*/Case::kKeep,