--- a/src/idl_gen_go.cpp
+++ b/src/idl_gen_go.cpp
@@ -73,6 +73,7 @@ static Namer::Config GoDefaultConfig() {
            /*namespace_seperator=*/"__",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kKeep,