--- a/src/idl_gen_kotlin.cpp
+++ b/src/idl_gen_kotlin.cpp
@@ -62,6 +62,7 @@ static Namer::Config KotlinDefaultConfig() {
            /*namespace_seperator=*/"__",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kKeep,