--- a/src/idl_gen_python.cpp
+++ b/src/idl_gen_python.cpp
@@ -61,6 +61,7 @@ static Namer::Config PythonDefaultConfig() {
            /*namespace_seperator=*/".",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kKeep,