--- a/src/idl_gen_rust.cpp
+++ b/src/idl_gen_rust.cpp
@@ -47,6 +47,7 @@ static Namer::Config RustDefaultConfig() {
            /*namespace_seperator=*/"::",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kSnake,