--- a/src/idl_gen_swift.cpp
+++ b/src/idl_gen_swift.cpp
@@ -45,6 +45,7 @@ static Namer::Config SwiftDefaultConfig() {
            /*namespace_seperator=*/"_",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kKeep,