--- a/src/idl_gen_ts.cpp
+++ b/src/idl_gen_ts.cpp
@@ -65,6 +65,7 @@ Namer::Config TypeScriptDefaultConfig() {
            /*namespace_seperator=*/"_",
            /*object_prefix=*/"",
            /*object_suffix=*/"T",
+           /*keywords_case_sensitive=*/true,
            /*keyword_prefix=*/"",
            /*keyword_suffix=*/"_",
            /*filenames=*/Case::kDasher,