--- a/src/idl_namer.h
+++ b/src/idl_namer.h
@@ -151,6 +151,16 @@ class IdlNamer : public Namer {
     return ConvertCase(d.name, config_.fields, Case::kLowerCamel) + "_type";
   }
 
+  // Php method for structs are a mix of keep case and lower camel case.
+  std::string LegacyPhpMethod(const StructDef &d,
+                              const std::string &suffix) const {
+    return d.name + suffix;
+  }
+  std::string LegacyPhpField(const std::string &prefix,
+                             const FieldDef &field) const {
+    return prefix + ConvertCase(field.name, Case::kUpperCamel);
+  }
+
  private:
   std::string NamespacedString(const struct Namespace *ns,
                                const std::string &str) const {