--- a/src/idl_namer.h
+++ b/src/idl_namer.h
@@ -154,7 +154,11 @@ class IdlNamer : public Namer {
   // Php method for structs are a mix of keep case and lower camel case.
   std::string LegacyPhpMethod(const StructDef &d,
                               const std::string &suffix) const {
-    return d.name + suffix;
+    return Type(d.name) + suffix;
+  }
+  std::string LegacyPhpMethod(const std::string &prefix,
+                              const StructDef &d) const {
+    return prefix + Type(d.name);
   }
   std::string LegacyPhpField(const std::string &prefix,
                              const FieldDef &field) const {