--- a/tests/MyGame/Example/StructOfStructs.php
+++ b/tests/MyGame/Example/StructOfStructs.php
@@ -7,6 +7,7 @@ use \Google\FlatBuffers\Struct;
 use \Google\FlatBuffers\Table;
 use \Google\FlatBuffers\ByteBuffer;
 use \Google\FlatBuffers\FlatBufferBuilder;
+use \Google\FlatBuffers\Constants;
 
 class StructOfStructs extends Struct
 {