--- a/tests/MyGame/Example2/Monster.php
+++ b/tests/MyGame/Example2/Monster.php
@@ -7,6 +7,7 @@ use \Google\FlatBuffers\Struct;
 use \Google\FlatBuffers\Table;
 use \Google\FlatBuffers\ByteBuffer;
 use \Google\FlatBuffers\FlatBufferBuilder;
+use \Google\FlatBuffers\Constants;
 
 class Monster extends Table
 {
@@ -17,6 +18,17 @@ class Monster extends Table
     public static function getRootAsMonster(ByteBuffer $bb)
     {
         $obj = new Monster();
+        return ($obj->init($bb->getInt($bb->getPosition()) + $bb->getPosition(), $bb));
+    }
+
+    /**
+     * @param ByteBuffer $bb
+     * @return Monster
+     */
+    public static function getSizePrefixedRootAsMonster(ByteBuffer $bb)
+    {
+        $obj = new Monster();
+        $bb->setPosition($bb->getPosition() + Constants::SIZEOF_INT);
         return ($obj->init($bb->getInt($bb->getPosition()) + $bb->getPosition(), $bb));
     }