@razvanalex 2024-10-08
Will close, as I will split this PR in smaller ones.